+++ test/integration/server_usage/post_process.py
@@ -21,16 +21,14 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import server_usage
-import lib.gen_libs as gen_libs
-import mongo_lib.mongo_libs as mongo_libs
-import mongo_lib.mongo_class as mongo_class
-import version
+import server_usage                             # pylint:disable=E0401,C0413
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class ArgParser(object):
+class ArgParser():                                      # pylint:disable=R0903
 
     """Class:  ArgParser
 
@@ -52,7 +50,7 @@ class ArgParser(object):
 
         """
 
-        self.args_array = dict()
+        self.args_array = {}
 
     def arg_exist(self, arg):
 
@@ -64,7 +62,7 @@ class ArgParser(object):
 
         """
 
-        return True if arg in self.args_array else False
+        return arg in self.args_array
 
 
 class UnitTest(unittest.TestCase):
@@ -77,7 +75,6 @@ class UnitTest(unittest.TestCase):
         setUp
         test_raw_print
         test_format_print
-        test_mongo_insert
         tearDown
 
     """
@@ -102,17 +99,6 @@ class UnitTest(unittest.TestCase):
         self.test_path = os.path.join(os.getcwd(), self.base_dir)
         self.config_path = os.path.join(self.test_path, "config")
         self.cfg = gen_libs.load_module("configuration", self.config_path)
-        svr = mongo_libs.create_instance(
-            "configuration", self.config_path, mongo_class.Server)
-        svr.connect()
-
-        if self.cfg.db in svr.fetch_dbs():
-            print("ERROR:  Test environment not clean - database: %s exists"
-                  % (self.cfg.db))
-            mongo_libs.disconnect([svr])
-            self.skipTest("Pre-conditions not met.")
-
-        mongo_libs.disconnect([svr])
 
     def test_raw_print(self):
 
@@ -139,38 +125,6 @@ class UnitTest(unittest.TestCase):
         """
 
         self.assertFalse(server_usage.post_process({}, self.args2, self.cfg))
-
-    def test_mongo_insert(self):
-
-        """Function:  test_mongo_insert
-
-        Description:  Test inserting data into Mongo database.
-
-        Arguments:
-
-        """
-
-        server_usage.post_process(self.proc_data, self.args3, self.cfg)
-        coll = mongo_libs.crt_coll_inst(self.cfg, self.cfg.db, self.cfg.coll)
-        coll.connect()
-
-        self.assertTrue(coll.coll_cnt() == 1)
-
-    def tearDown(self):
-
-        """Function:  tearDown
-
-        Description:  Clean up of integration testing.
-
-        Arguments:
-
-        """
-
-        mongo = mongo_libs.create_instance(
-            "configuration", self.config_path, mongo_class.DB)
-        mongo.db_connect(self.cfg.db)
-        mongo.db_cmd("dropDatabase")
-        mongo_libs.disconnect([mongo])
 
 
 if __name__ == "__main__":