+++ test/integration/server_usage/run_program.py
@@ -21,16 +21,14 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import server_usage
-import lib.gen_libs as gen_libs
-import mongo_lib.mongo_libs as mongo_libs
-import mongo_lib.mongo_class as mongo_class
-import version
+import server_usage                             # pylint:disable=E0401,C0413
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class ArgParser(object):
+class ArgParser():                                      # pylint:disable=R0903
 
     """Class:  ArgParser
 
@@ -53,7 +51,7 @@ class ArgParser(object):
         """
 
         self.cmdline = None
-        self.args = dict()
+        self.args = {}
 
     def get_val(self, skey, def_val=None):
 
@@ -76,11 +74,9 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_mongo
         test_print_format
         test_no_print
         test_print_raw
-        tearDown
 
     """
 
@@ -104,34 +100,6 @@ class UnitTest(unittest.TestCase):
         self.args2.args_array = {
             "-c": "configuration", "-d": self.config_path, "-n": True,
             "-m": True}
-        svr = mongo_libs.create_instance(
-            "configuration", self.config_path, mongo_class.Server)
-        svr.connect()
-
-        if self.cfg.db in svr.fetch_dbs():
-            print("ERROR:  Test environment not clean - database: %s exists"
-                  % (self.cfg.db))
-            mongo_libs.disconnect([svr])
-            self.skipTest("Pre-conditions not met.")
-
-        mongo_libs.disconnect([svr])
-
-    def test_mongo(self):
-
-        """Function:  test_mongo
-
-        Description:  Test inserting data into Mongo database.
-
-        Arguments:
-
-        """
-
-        server_usage.run_program(self.args2)
-
-        coll = mongo_libs.crt_coll_inst(self.cfg, self.cfg.db, self.cfg.coll)
-        coll.connect()
-
-        self.assertTrue(coll.coll_cnt() == 1)
 
     def test_print_format(self):
 
@@ -172,22 +140,6 @@ class UnitTest(unittest.TestCase):
 
         with gen_libs.no_std_out():
             self.assertFalse(server_usage.run_program(self.args))
-
-    def tearDown(self):
-
-        """Function:  tearDown
-
-        Description:  Clean up of integration testing.
-
-        Arguments:
-
-        """
-
-        mongo = mongo_libs.create_instance(
-            "configuration", self.config_path, mongo_class.DB)
-        mongo.db_connect(self.cfg.db)
-        mongo.db_cmd("dropDatabase")
-        mongo_libs.disconnect([mongo])
 
 
 if __name__ == "__main__":